refactor the parsing logic and move to cli/starters.py from tepmplate
