+++ kedro/templates/project/hooks/utils.py
@@ -47,48 +47,12 @@ docs = [
 ]
 """
 
-def _validate_range(start, end):
-    if int(start) > int(end):
-        message = f"'{start}-{end}' is an invalid range for project add-ons.\nPlease ensure range values go from smaller to larger."
-        click.secho(message, fg="red", err=True)
-        sys.exit(1)
-
-def _validate_selection(add_ons):
-    for add_on in add_ons:
-        if int(add_on) < 1 or int(add_on) > 5:
-            message = f"'{add_on}' is not a valid selection.\nPlease select from the available add-ons: 1, 2, 3, 4, 5."
-            click.secho(message, fg="red", err=True)
-            sys.exit(1)
 
 
-def parse_add_ons_input(add_ons_str):
-    """Parse the add-ons input string.
 
-    Args:
-        add_ons_str: Input string from prompts.yml.
 
-    Returns:
-        list: List of selected add-ons as strings.
-    """
-    if add_ons_str == "all":
-        return ["1", "2", "3", "4", "5"]
-    if add_ons_str == "none":
-        return []
 
-    # Split by comma
-    add_ons_choices = add_ons_str.split(",")
-    selected = []
 
-    for choice in add_ons_choices:
-        if "-" in choice:
-            start, end = choice.split("-")
-            _validate_range(start, end)
-            selected.extend(str(i) for i in range(int(start), int(end) + 1))
-        else:
-            selected.append(choice.strip())
-
-    _validate_selection(selected)
-    return selected
 
 
 def setup_template_add_ons(selected_add_ons_list, requirements_file_path, pyproject_file_path):
