+++ kedro/templates/project/{{ cookiecutter.repo_name }}/pyproject.toml
@@ -37,3 +37,4 @@ namespaces = false
 package_name = "{{ cookiecutter.python_package }}"
 project_name = "{{ cookiecutter.project_name }}"
 kedro_init_version = "{{ cookiecutter.kedro_version }}"
+add_ons = [{{ cookiecutter.add_ons }}]
